formatting test module
